Split resolver in a different class
